requirements.txt
#Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1

# Async HTTP
aiohttp==3.9.5
httpx==0.27.0

#Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

#Cache
redis[hiredis]==5.0.4

#Kafka
aiokafka==0.10.0

# AI / LLM
anthropic==0.26.0
openai==1.30.0
langchain==0.2.0
langgraph==0.0.62
tiktoken==0.7.0

#Vector store
weaviate-client==4.6.1

#Security scanning(Python wrappers)
bandit==1.7.8
semgrep==1.73.0

# Observability
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
prometheus-client==0.20.0
structlog==24.1.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
tenacity==8.3.0
rich==13.7.1
typer==0.12.3
